modules/api/location_consumer/requirements.txt
Flask==1.1.1
Flask_Cors==3.0.8
Flask_SQLAlchemy==2.4.0
GeoAlchemy2==0.8.4
kafka_python==2.0.2
marshmallow==3.7.1
marshmallow_sqlalchemy==0.23.1
SQLAlchemy==1.3.19
MarkupSafe==1.1.1
Werkzeug==0.16.1
itsdangerous==1.1.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
shapely==1.7.0
